requirements.txt
flake8
coverage
snakeviz
memory_profiler
sphinx
sphinx_rtd_theme
setuptools
wheel
gitpython
